requirements.txt
asgiref==3.4.1
boto3==1.16.63
botocore==1.19.63
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.1
jmespath==0.10.0
Markdown==3.3.6
numpy==1.22.3
pandas==1.4.1
Pillow==9.1.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.3.7
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.9
waitress==2.0.0
whitenoise==5.3.0
zipp==3.7.0
